Testing a new approach for this.
First approach of doing everything at compile time was a bad idea and would
result in unflexible bindings since nobody want to recompile a program just
to add support for a new language.

New approach reads the tree-sitter CLI configuration file and dynamic loads
the parsers found, so in theory you just need to have the parsers visible to
tree-sitter CLI and it must work.

This also means that at least JSON and C (for now) parsers must be installed
for the unit tests to pass.